Initial Commit
I've decided to take the leap with Rails 5.1's built in support for
webpack and build the application in a decoupled front-end style with
react. This will allow the front-end and back-end code to vary
independently, allow font-end developers to use industry standard
tooling. The trade-off being a bit of increased complexity in the client
app dealing with data, and bit of complexity managing building and
running two apps at once in development and test environments. I've
decided to use foreman to manage the latter.

Tech Stack:
 * Rails version 5.1 (pre-release)
 * React Front-end
 * RSpec and Capybara for testing
 * Postgresql database