Teachers Can Create Course
I created the simplest path through the app that I could think of in
creating a course with just a name. Eventually we will want to be able
to edit, and delete courses. We will also need to be able to give
students grades. Working on the teacher stories seems like it should be
a good approach to getting most of the app bootstrapped.